+++ src/firebase/database.js
@@ -1,0 +1,11 @@
+import { 
+  getDatabase,
+  ref,
+  onValue,
+  push,
+  set
+} from "firebase/database";
+import { app } from "./firebaseConfig";
+
+
+export const db = getDatabase(app);
